refactor: extract useMediaItems composable from VideoDialog (3b)
videoMedia/subtitleMedia state, caption/subtitle URL computeds, the
media-item fetches, and the selected-video/language watchers move to
composables/useMediaItems.ts; the player captures its resume position
via the onBeforeLanguageReload callback.
